ARM GCC defaults char to unsigned.  Declaring sync as signed explicitly, and removing the cast of "-1" to char. This is to ensure correct behavior based on language standards:
"If the destination type is unsigned, the resulting value is the least unsigned integer congruent to the source integer (modulo 2^n where n is the number of bits used to represent the unsigned type). [ Note: In a two’s complement representation, this conversion is conceptual and there is no change in the bit pattern (if there is no truncation). —end note ]

Former-commit-id: f5b1354bf8e80a3c2ff95123f177029df8e7cb80